updates viewport size detection to more reliable method
Per http://www.w3schools.com/js/js_window.asp , the value
document.body.clientHeight is the size of the page, not the viewport,
if the page is smaller than viewport (rare but possible). This is not
what we want, as we are detecting screen size or screen edges. Instead
window.innerHeight works on browsers newer than IE8.