Replace .hoverable with .no-hover [#117179271]
The original assumption was that all .dropdown-wrapper elements would
respond to the :hover event.

A recent PR required some .dropdown-wrapper elements to *only* respond
to click events (not :hover). It introduced the .hoverable class as a
means to denone those .dropdown-wrapper elements that should respond to
the hover event, and added click handlers for the specific elements that
should respond to click events. This changed the basic assumption such
that .dropdown-wrapper elements had *no* default behavior and had to
explicitly be .hoverable or have a click handler written.

This commit removes the functionality conveyed by the .hoverable class
and restores the original assumption that all .dropdown-wrapper elements
respond to the :hover event. It then introduces the .no-hover class to
explicitly prevent this default behavior.