[#122089131] Use separate logic check for geolocation rules (#1032)
* [#122089131] Use separate logic check for geolocation rules

Geolocation data is undefined on page-load. It’s populated after the
geolocater query finishes and we re-run the logic that decides which
bars to show on the page. We had been showing site elements where
country != viewer’s country and “viewer’s country” was undefined. Our
!= logic did not take into account that all geolocation data is
undefined on page load - so when the page did load, we’d already shown
the site element even if the viewer was in the country where the site
element was not supposed to show up

* Refactor and unit test