Simplify themes feature spec
This spec is very error prone on CircleCI (and to some degree locally as
well) due to the amount of stuff it does. It basically clicks through
about 10 different themes and checks whether selecting each theme
updates the background color in the UI.

This is too much. It is enough to say that clicking only on one theme
updates the background color UI to verify that the feature works, so
that's what I did.