Make topbars respect visibility cookies
When there is a 'Dismissed' cookie being present, a topbar will still be
shown, but in a 'stay-hidden' state (i.e. there will be only a small
"pull-down" arrow visible, which will show the bar when click on).

I don't believe there is a need for a logic to respect/remove the
'Success' cookie as the topbars don't disappear on their own on success.