Working GetResponse tagging
Due to the fact that you cannot assign tags to a contact during its
creation via API at GetResponse (their API v3 does not support that,
see: https://apidocs.getresponse.com/v3/resources/contacts#contacts.create)
and because you cannot fetch a contact via API if it has not yet been
confirmed (i.e. for double optin campaigns/lists at GetResponse, you
need to first "Subscribe" to the list, which is basically email
confirmation) tagging is implemented in a different way.

For each newly added subscriber, we fetch two most recent confirmed
contacts from the appropriate list/campaign and we assign chosen tags to
those two contacts. Two, because there could be race conditions, which
this should eliminate. It will tag most contacts twice, but it is OK,
GetResponse is not complaining if you assign the same set of tags more
than once.

This is a bit of a hack, but it gets the end-user functionality
implemented. Not 100%, but I think it's 95% there.