Keep only 15 latest releases while deploying
We were running into the "no space left on the device" errors; the disk
was simply running out of free inodes; each release is about 46K files;
we were keeping 50 last releases, which gives about 2,3mln inodes, while
the limit on the disk is about 2mln. That’s why the errors.