Nginx: do not proxy_pass on timeouts
Basically if there is a timeout in one of the upstream processes (thin),
don't pass this request to the next upstram process. Just timeout.

This is so that when we process large animated gif file uploads (which
take a couple of minutes to process), we don't send subsequent file
upload requests, which at this point is like a small DoS attack as the
server is overloaded with resizing the gif into multiple sizes.