Don't regenrate site script on ContactList update
Original idea was to invalidate Site's cache when we subscribe a
contact. This is a good idea as this would make all site stats
(views/conversions) always be freshen up on new subscriptions (but not
on new views). However, it would also mean that the site script would be
regenerated on each subscription.

That's too much.

Cache longer (display more outdated data) instead.