Allow users to create sites with the same url
And also allow users to have membership access to sites with the same
url.

We had a unique url validation in place, which checked if the site that
you are adding has a unique url among all your sites and among other
users who have membership access to this particular site. In theory this
works well, however in practice, this proves to be problematic. Consider
the following situation:

* User A has site eventbrite.com
* User B also has a site eventbrite.com

As long as those users/sites are not connected with each other,
everything is fine. However, things break when user A invites user B to
have access to his eventbrite.com site. Suddenly for user A both his
eventbrite.com sites are invalid.  The solution we think is appropriate
to solve this is to just let end-users figure this out. In other words,
we're going to drop this unique url validation and if users have 10
sites in their list, all with the same eventbrite.com url, then all
good. It would be upon users to clean this up (i.e. delete or rename
those duplicate sites).