Refactor DynamoDB to handle pagination
DynamoDB#query has been changed in order to handle DynamoDB pagination.
AWS has a limitation of 1 MB for query response. If resonse exceeds this
number, the results are divided into "pages" of data that are 1 MB each.

In order to provide seamless integration it accepts a block to iterate
over response items and returns enumeration if no block given.