Increase Shoryuken threads to 10
We've had an issue with Maropost taking a lot of time (up to 40
minutes!) to sync contacts. For a busy list conneted to Maropost this
meant saturating our workers (I'm looking at you draxe.com). Increasing
it to 10 should not be that bad when it comes to site script
regeneration / weekly email digests, but we'll just have to see.